organise code into routes and controllers
