+++ collaboration service/routes/sessionRoutes.js
@@ -1,0 +1,22 @@
+// routes/sessionRoutes.js
+
+const express = require('express');
+const router = express.Router();
+const sessionController = require('../controllers/sessionController');
+
+// This route serves the socket connection for session handling
+module.exports = (io) => {
+  io.on('connection', (socket) => {
+    console.log('A user connected:', socket.id);
+
+    // Delegate session handling to the controller
+    sessionController.joinSession(socket, io);
+
+    // Handle disconnect
+    socket.on('disconnect', () => {
+      console.log('User disconnected:', socket.id);
+    });
+  });
+
+  return router;
+};
